Added changes to create react app
