+++ potluck-planner/src/Components/Login.js
@@ -1,0 +1,13 @@
+
+
+
+export default function Login(props) {
+
+    return(
+        <div> 
+            <h1>Hello</h1>
+
+        </div>
+
+    );  
+  }
